Various Windows related cleanups.
This modifies some of the test infrastructure so that more tests can
work on Windows. Specifically, we can now write tests that use symlinks
and still run on Windows.

We also clean up the implementation of `is_same_file` by reusing
types/functions from the kernel32 and winapi crates. Also, add a few
tests for this function.