errors: keep context when converting to io::Error
This commit tweaks the `From<walkdir::Error> for io::Error`
implementation to always retain the current context when
constructing the `io::Error`. This differs from the previous
implementation in that the original raw I/O error is no longer
returned.

To compensate, a new method, `into_io_error`, has been
added which returns the original I/O error, if one exists.

We do not consider this a breaking change because the
documentation for the `From` impl always stated that it
existed for ergonomic reasons. Arguably, the implementation
in this commit is a more faithful reflection of that
documentation.

This commit also clears up the public documentation
surrounding the aforementioned methods.