windows: fix OneDrive traversals
This commit fixes a bug on Windows where walkdir refused to traverse
directories that resided on OneDrive via its "file on demand" strategy.
The specific bug is that Rust's standard library treats a reparse point
(which is what OneDrive uses) as distinct from a file or directory, which
wreaks havoc on any code that uses FileType::{is_file, is_dir}. We fix
this by checking the directory status of a file by looking only at whether
its directory bit is set.

This bug was originally reported in ripgrep:
https://github.com/BurntSushi/ripgrep/issues/705

It has also been filed upstream:
https://github.com/rust-lang/rust/issues/46484

And has a pending fix:
https://github.com/rust-lang/rust/pull/47956