tests: rewrite them
This gets rid of a lot of unnecessary infrastructure around maintaining
the directory hierarchy in a tree. This was principally used in order to
support effective quickcheck tests, but since we dropped quickcheck, we
no longer need such things.

We know center tests around the Dir type, which makes setting up the
tests simpler and easier to understand.