tests: add contents_first with min & max depth
This just adds `contents_first(true)` versions of the existing tests for
min and max depth. Given that handling contents first can be a bit
confusing, it seemed wise to test it more thoroughly.

For symmetry, this switches the existing min and max depth tests to use
`path()` instead of `sorted_paths()`. Because all of the test
directories only have one entry, there should never be an issue with
entries being returned in a different, but still correct, order.