bug: filter_entry misbehaved when contents_first is set
The behavior of `IntoIter::filter_entry` contradicted its documentation
when the `contents_first` was enabled and the predicate returned `false`
for directories.

> Note that if the iterator has contents_first enabled, then this method
> is no different than calling the standard Iterator::filter method

According to the documentation, it must list all the entries in the
skipped directories anyway, but listed only few of them.

This commit fixes the issue by making `FilterEntry::next` call
a newly introduced `IntoIter::skip_current_dir_unless_contents_first`
method instead of the `IntoIter::skip_current_dir` method when the
predicate returns `false` for a directory. `FilterEntry::next` used to
skip a wrong directory when the predicate returned `false`.